Delete PlatformAccessMethod

+++ grails-app/migrations/changelogs/2023-04-14.groovy
@@ -2,10 +2,11 @@ package changelogs
 
 databaseChangeLog = {
 
+
     changeSet(author: "djebeniani (generated)", id: "1681469132222-1") {
         grailsChange {
             change {
-                sql.execute("delete from org_access_point where oar_access_method_rv_fk = (select rdv_id from refdata_value where rdv_value = 'vpn')")
+                sql.execute("delete from access_point_data where apd_org_access_point_fk in (select oar_id from org_access_point where oar_access_method_rv_fk = (select rdv_id from refdata_value where rdv_value = 'vpn'))")
             }
             rollback {}
         }
@@ -14,7 +15,7 @@ databaseChangeLog = {
     changeSet(author: "djebeniani (generated)", id: "1681469132222-2") {
         grailsChange {
             change {
-                sql.execute("delete from access_point_data where apd_org_access_point_fk in (select oar_id from org_access_point where oar_access_method_rv_fk = (select rdv_id from refdata_value where rdv_value = 'vpn'))")
+                sql.execute("delete from org_access_point where oar_access_method_rv_fk = (select rdv_id from refdata_value where rdv_value = 'vpn')")
             }
             rollback {}
         }
